Assignment 2 final
Added additional requirements and changed a couple:
More cleanup.

I took a div class out of the <article> section so it the flex requirements do not alter the content.  Not sure if that's acceptable for this assignment but it now looks more like the wire.  The 'container' class is part of the body but I wanted it independent from the 3 sections.

Also I changed the 'get a piece of mind' color to green because the required black text overlaps with the iphone.  Hopefully thats fine, especially when it looks closer to the wire.

I'm too tired to try to figure out how to remove the white borders on the top and bottom so a solution would be awesome.