not much changes
the only thing i done is add a hash to the button links..  the horizontal white border bars are killing me though so how would i remove those?

also i was chasing my tail on this placement of the objects against the image background for the header.  i tried to make the header area a flex object but when i did my button was stretched to the width of the background... i couldn't figure out why and it was due to the background (i placed the button near the blue button to put it in a different nest and it looked normal).  anyways, i may be down to discuss with you this thursday office hours =)