Lotsa Moving Around and Emergency Patchwork
So I tried to intermingle the landing page assignment with the blog and I epic failed :-(... As a temporary fix I just had both pages read from separate css stylesheets though it probably will hurt my grade.  Couple things from the last homework assignment I'm not quite clear on:

When to use the reset.css and normalize.css (as it relates to this project).  Is it a template to begin with or something to save over from time to time to clean up whatever while building out the layout?

As it relates to intermingling both pages on the same css, how do I use 2 different grid styles without them affecting each other's layout?  If I had more time I would try to figure this problem out.  Im sure the hint is in elements, id, class { grid #1} but I probably need to talk to you in person about this.